requirements.txt
httpx==0.27.0
pandas==2.2.2
langchain==0.2.11
langgraph==0.1.19
aiogram==3.10.0
python-dotenv==1.0.1
langchain_openai==0.1.20
langchainhub==0.1.20
tabulate==0.9.0